feat(iota-types): move duplicated `estimate_gas_budget_from_gas_cost` to `iota-types`
