+++ crates/iota/src/displays/dry_run_tx_block.rs
@@ -8,12 +8,13 @@ use iota_json_rpc_types::{
     DryRunTransactionBlockResponse, IotaTransactionBlockDataAPI, IotaTransactionBlockEffectsAPI,
     ObjectChange,
 };
+use iota_types::gas::estimate_gas_budget_from_gas_cost;
 use tabled::{
     builder::Builder as TableBuilder,
     settings::{Panel as TablePanel, Style as TableStyle, style::HorizontalLine},
 };
 
-use crate::{client_commands::estimate_gas_budget_from_gas_cost, displays::Pretty};
+use crate::displays::Pretty;
 impl Display for Pretty<'_, DryRunTransactionBlockResponse> {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         let Pretty(response) = self;
